Added Use Case for settings Made temperature unit selection functional - Using settings use case, selected temperature unit is propagated from viewModel down to repository transformer - Added transformer function to convert temperature based on the selected temperature unit

+++ app/src/main/java/com/weather/nimbus/data/weather/model/ForecastData.kt
@@ -17,7 +17,7 @@ data class ForecastData(
         val date: Int,
         val dayOfWeek: String,
         val weatherStatus: WeatherStatus,
-        val minTemperature: Double,
-        val maxTemperature: Double
+        val minTemperature: Int,
+        val maxTemperature: Int
     )
 }
